+++ src/main/java/com/monginis/ops/model/PostFrItemStockDetail.java
@@ -8,148 +8,117 @@ public class PostFrItemStockDetail {
 
 	private String itemId;
 
-	private int regOpeningStock;
+	private float regOpeningStock;
 
-	private int spOpeningStock;
+	private float spOpeningStock;
 
-	private int physicalStock;
+	private float physicalStock;
 
-	private int stockDifference;
+	private float stockDifference;
 
-	private int regTotalPurchase;
+	private float regTotalPurchase;
 
-	private int spTotalPurchase;
+	private float spTotalPurchase;
 
-	private int regTotalGrnGvn;
+	private float regTotalGrnGvn;
 
-	private int regTotalSell;
+	private float regTotalSell;
 
-	private int spTotalSell;
+	private float spTotalSell;
 
 	private String remark;
 	private String itemName;
 	private String itemCode;
-
 	public int getOpeningStockDetailId() {
 		return openingStockDetailId;
 	}
-
 	public void setOpeningStockDetailId(int openingStockDetailId) {
 		this.openingStockDetailId = openingStockDetailId;
 	}
-
 	public int getOpeningStockHeaderId() {
 		return openingStockHeaderId;
 	}
-
 	public void setOpeningStockHeaderId(int openingStockHeaderId) {
 		this.openingStockHeaderId = openingStockHeaderId;
 	}
-
 	public String getItemId() {
 		return itemId;
 	}
-
 	public void setItemId(String itemId) {
 		this.itemId = itemId;
 	}
-
-	public int getRegOpeningStock() {
+	public float getRegOpeningStock() {
 		return regOpeningStock;
 	}
-
-	public void setRegOpeningStock(int regOpeningStock) {
+	public void setRegOpeningStock(float regOpeningStock) {
 		this.regOpeningStock = regOpeningStock;
 	}
-
-	public int getSpOpeningStock() {
+	public float getSpOpeningStock() {
 		return spOpeningStock;
 	}
-
-	public void setSpOpeningStock(int spOpeningStock) {
+	public void setSpOpeningStock(float spOpeningStock) {
 		this.spOpeningStock = spOpeningStock;
 	}
-
-	public int getPhysicalStock() {
+	public float getPhysicalStock() {
 		return physicalStock;
 	}
-
-	public void setPhysicalStock(int physicalStock) {
+	public void setPhysicalStock(float physicalStock) {
 		this.physicalStock = physicalStock;
 	}
-
-	public int getStockDifference() {
+	public float getStockDifference() {
 		return stockDifference;
 	}
-
-	public void setStockDifference(int stockDifference) {
+	public void setStockDifference(float stockDifference) {
 		this.stockDifference = stockDifference;
 	}
-
-	public int getRegTotalPurchase() {
+	public float getRegTotalPurchase() {
 		return regTotalPurchase;
 	}
-
-	public void setRegTotalPurchase(int regTotalPurchase) {
+	public void setRegTotalPurchase(float regTotalPurchase) {
 		this.regTotalPurchase = regTotalPurchase;
 	}
-
-	public int getSpTotalPurchase() {
+	public float getSpTotalPurchase() {
 		return spTotalPurchase;
 	}
-
-	public void setSpTotalPurchase(int spTotalPurchase) {
+	public void setSpTotalPurchase(float spTotalPurchase) {
 		this.spTotalPurchase = spTotalPurchase;
 	}
-
-	public int getRegTotalGrnGvn() {
+	public float getRegTotalGrnGvn() {
 		return regTotalGrnGvn;
 	}
-
-	public void setRegTotalGrnGvn(int regTotalGrnGvn) {
+	public void setRegTotalGrnGvn(float regTotalGrnGvn) {
 		this.regTotalGrnGvn = regTotalGrnGvn;
 	}
-
-	public int getRegTotalSell() {
+	public float getRegTotalSell() {
 		return regTotalSell;
 	}
-
-	public void setRegTotalSell(int regTotalSell) {
+	public void setRegTotalSell(float regTotalSell) {
 		this.regTotalSell = regTotalSell;
 	}
-
-	public int getSpTotalSell() {
+	public float getSpTotalSell() {
 		return spTotalSell;
 	}
-
-	public void setSpTotalSell(int spTotalSell) {
+	public void setSpTotalSell(float spTotalSell) {
 		this.spTotalSell = spTotalSell;
 	}
-
 	public String getRemark() {
 		return remark;
 	}
-
 	public void setRemark(String remark) {
 		this.remark = remark;
 	}
-
 	public String getItemName() {
 		return itemName;
 	}
-
 	public void setItemName(String itemName) {
 		this.itemName = itemName;
 	}
-
 	public String getItemCode() {
 		return itemCode;
 	}
-
 	public void setItemCode(String itemCode) {
 		this.itemCode = itemCode;
 	}
-
 	@Override
 	public String toString() {
 		return "PostFrItemStockDetail [openingStockDetailId=" + openingStockDetailId + ", openingStockHeaderId="
@@ -160,4 +129,5 @@ public class PostFrItemStockDetail {
 				+ spTotalSell + ", remark=" + remark + ", itemName=" + itemName + ", itemCode=" + itemCode + "]";
 	}
 
+	
 }
